+++ packages/media/media-card/src/root/card/index.tsx
@@ -315,13 +315,24 @@ export class Card extends Component<CardProps, CardState> {
       this.setState({
         isPlayingFile: true,
       });
-    } else if (shouldOpenMediaViewer && identifier.mediaItemType === 'file') {
-      const mediaViewerSelectedItem: FileIdentifier = {
-        id: await identifier.id,
-        mediaItemType: 'file',
-        collectionName: identifier.collectionName,
-        occurrenceKey: identifier.occurrenceKey,
-      };
+    } else if (shouldOpenMediaViewer) {
+      let mediaViewerSelectedItem: Identifier | undefined;
+
+      if (isFileIdentifier(identifier)) {
+        mediaViewerSelectedItem = {
+          id: await identifier.id,
+          mediaItemType: 'file',
+          collectionName: identifier.collectionName,
+          occurrenceKey: identifier.occurrenceKey,
+        };
+      } else {
+        mediaViewerSelectedItem = {
+          mediaItemType: 'external-image',
+          dataURI: identifier.dataURI,
+          name: identifier.name,
+        };
+      }
+
       this.setState({
         mediaViewerSelectedItem,
       });
@@ -362,54 +373,19 @@ export class Card extends Component<CardProps, CardState> {
     });
   };
 
-  // returns a valid MV data source including current the card identifier
-  getMediaViewerDataSource = (): MediaViewerDataSource => {
-    const { mediaViewerDataSource } = this.props;
-    const { mediaViewerSelectedItem } = this.state;
-
-    if (!mediaViewerSelectedItem) {
-      return {
-        list: [],
-      };
-    }
-
-    if (!mediaViewerDataSource) {
-      return {
-        list: [mediaViewerSelectedItem],
-      };
-    }
-
-    // we want to ensure the card identifier is in the list
-    const { list } = mediaViewerDataSource;
-    if (
-      list &&
-      list.length &&
-      mediaViewerSelectedItem.mediaItemType === 'file'
-    ) {
-      const isSelectedItemInList = list.some(
-        item =>
-          item.mediaItemType === 'file' &&
-          item.id === mediaViewerSelectedItem.id,
-      );
-      if (!isSelectedItemInList) {
-        return {
-          list: [mediaViewerSelectedItem, ...list],
-        };
-      }
-    }
-
-    return mediaViewerDataSource;
-  };
-
   renderMediaViewer = () => {
     const { mediaViewerSelectedItem } = this.state;
-    const { context, identifier } = this.props;
-    if (!mediaViewerSelectedItem || identifier.mediaItemType !== 'file') {
+    const { context, identifier, mediaViewerDataSource } = this.props;
+    if (!mediaViewerSelectedItem) {
       return;
     }
 
-    const { collectionName = '' } = identifier;
-    const dataSource = this.getMediaViewerDataSource();
+    const collectionName = isFileIdentifier(identifier)
+      ? identifier.collectionName || ''
+      : '';
+    const dataSource: MediaViewerDataSource = mediaViewerDataSource || {
+      list: [],
+    };
 
     return ReactDOM.createPortal(
       <MediaViewer